Fixed the bug that wouldn't allow negative values.
Needed a set point buffer to retain the negative value sent to the controller in order to allow for negative torques. Also fixed up the PID loop to allow for large input values. It is important to not go too far over the rated current though since we can easily burn out the motor.